OO-111: added parameter mapping for more than 23 params. Adjusted update and insert methods to store modification flags
