+++ src/main/scala/ch/openolitor/core/repositories/BaseParameter.scala
@@ -1,0 +1,5 @@
+package ch.openolitor.core.repositories
+
+trait BaseParameter {
+  def parameter[V](value: V)(implicit binder: SqlBinder[V]): Any = binder.apply(value)
+}
