+++ src/main/scala/ch/openolitor/core/scalax/Tuple24.scala
@@ -20,14 +20,11 @@
 * with this program. If not, see http://www.gnu.org/licenses/                 *
 *                                                                             *
 \*                                                                           */
-package ch.openolitor.core.domain
+package ch.openolitor.core.scalax
 
-import ch.openolitor.core.models.UserId
-import org.joda.time.DateTime
-
-case class EventMetadata(originator: UserId, version: Int, timestamp: DateTime, seqNr: Long, source: String)
-
-trait PersistetEvent extends Serializable {
-  val meta: EventMetadata
+case class Tuple24[+T1, +T2, +T3, +T4, +T5, +T6, +T7, +T8, +T9, +T10, +T11, +T12, +T13, +T14, +T15, +T16, +T17, +T18, +T19, +T20, +T21, +T22, +T23, +T24](val _1: T1, val _2: T2, val _3: T3, val _4: T4, val _5: T5, val _6: T6, val _7: T7, val _8: T8, val _9: T9, val _10: T10, val _11: T11, val _12: T12, val _13: T13, val _14: T14, val _15: T15, val _16: T16, val _17: T17, val _18: T18, val _19: T19, val _20: T20, val _21: T21, val _22: T22, val _23: T23, val _24: T24)
+  extends Product24[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23, T24]
+  with CustomTuple {
+  override def canEqual(other: Any) =
+    other.isInstanceOf[Tuple24[_, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _]]
 }
-
