+++ src/main/scala/ch/openolitor/stammdaten/models/ProduzentenModels.scala
@@ -23,10 +23,11 @@
 package ch.openolitor.stammdaten.models
 
 import java.util.UUID
+
 import ch.openolitor.core.models._
 import java.util.Date
 import org.joda.time.DateTime
-import ch.openolitor.core.repositories.Product23
+import ch.openolitor.core.scalax.Product23
 
 case class BaseProduzentId(id: String)
